Classes Organized & Death State
+ Death State completed (Animation)
+ C++ Classes organized into folders
