+++ Content/Game/Blueprints/AI/Enemy/BP_EnemyCharacter.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:5f833699d7588a141f39fd29961d758807e653fb023319650ad65121222fbeda
-size 34872
+oid sha256:4679796214e1ccf483a6d31f593ab24c9dcfd899e356a4af4fa2a46770c9ef8f
+size 69499
