+++ Content/Game/Characters/Enemy/AI/ST_Enemy.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:890c6f73a8b7630db8fd590fa6afd2e945c78baba7484bb178df4e679ba11775
-size 23795
+oid sha256:44bd3cec185494bc50cc1b9fc66fc0383307b6b82961a977415f1da7e88ed2a2
+size 34445
