+++ Content/__ExternalActors__/Game/Levels/TestLevel/5/T6/PRHIAP17TGWDVTGO72O3YZ.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:a3ef64784bb0c0f9fcae1060d4a3dd0683ae1427f6b1ff141043d48a197b92aa
-size 6423
+oid sha256:a559ddd64fbca899e33a0079fcd6cdeb9178dc0b177b953ce8f969e1e75cd95e
+size 6648
