+++ Source/Ghost_SpecOps/Civilian/CivilianAIController.cpp
