+++ Source/Ghost_SpecOps/Civilian/CivilianAIController.h
