+++ Source/Ghost_SpecOps/Civilian/CivilianCharacter.cpp
