+++ Source/Ghost_SpecOps/Tasks/FindNextPatrolPoint.h
