+++ Source/Ghost_SpecOps/Tasks/FindRandomPointNearby.cpp
