+++ Source/Ghost_SpecOps/Tasks/FindRandomPointNearby.h
