+++ Source/Ghost_SpecOps/Tasks/FindSmartObject.cpp
