+++ Source/Ghost_SpecOps/Tasks/FindSmartObject.h
