+++ Source/Ghost_SpecOps/Tasks/PatrolPath.h
