+++ Source/Ghost_SpecOps/Tasks/UseSmartObject.cpp
