+++ Source/Ghost_SpecOps/Tasks/UseSmartObject.h
